Multiple Wells Affect One Target
Target object can now be affected by multiple wells. When the well can affect an object, the object adds it a list which checks if is currently affecting it. The well object adds the intended velocity that is should output to another list on the target, the target then adds all velocity’s on the list and calculates its new velocity.