Interactable Objects and Tid Bits
The player can now interact with objects by pressing 'e' while looking at objects that have an Interactable bool ticked, when an item is picked up the object hovers and follows the front of the players camera.
Gravity Wells now change colour depending on if they are attracting or repelling.
Create a room so that objects stop falling off when testing.
Changed the colour of the flooring to make objects easer to see.
Removed unnecessary scripts.