Gun Control and Objects
Pushed wrong repo with this description so adding to this description.

02/11/17 13:00
Implemented a basic UI show where the gun will fire a repulsion or attraction well, the power of the well and the cycle of the wells which removes the first well activated when there a 4th well is needed.
Added a capsule on the gun to indicate the power of the gun, the capsule scales depending on the power adjusted by the user.
When the user holds the fire button a preview of where a gravity well will land and what it will affect. A transparent sphere shows the radius and the colour show the polarity.
The gravity well is now launched on release of the fire button.
Picking up objects is now easer as it is on button down instead of the axis system.

02/11/17 23:00
Added two types of buttons, the first is an object button that is activate when an object or player is placed on it. The second type is a player button that can only be activated by the user interacting with it. Buttons can have one target which currently can be a door or a dispenser. Player buttons and object buttons both inherit a couple of methods from a button class.
Dispensers are activated by buttons and when activate they drop a cube into the world, if the dispenser is activated again the last cube it made is removed and a new one is created.
Doors are activated by buttons and can be toggled open or closed. A simple door is animated with an open and close animation.
Change the player movement so that it moves smoother by changing from adjusting the rigidbodys velocity to translating the movement. Jumping remains adjusting the players y velocity.