New Gravity Well Code and UI
Recoded the gravity wells so that they no longer rely on collision triggers as this caused some issues when and object didn’t leave a well properly.
The new code has a controller which checks if an object is in range of a well and tells the well if an object is, the gravity well then handles it as it would have before.

Gravity wells are now displayed on the UI showing the player what polarity each well is. When a player looks at a well the UI for that well is highlighted with a yellow boarder.
UI for the menu when playing has been put in place with a “quit game” and “quit to menu”.
Added a new scene for the main menu which has a “continue”, “new game”, “load game” and “quit to desktop”. “Continue” and “new game” currently do nothing as there is no infrastructure for it, however “load game” opens a window that the player can scroll through to choose a level but for the same reason they cannot load a level.