Object Prediction and Save/Load
When an object is held still in a gravity well the object now creates and invisible cube with the same transform as the object that leaves a train behind a trail simulating where the object would go, if the object is moved then the projection is destroyed.
Using serialisation, the players progress is now saved. On first run a file is created which stores the highest level and any loads of the game after will load this file. When the game is loaded a game object is created which is not destroyed between scenes which stores the player progress, this game object stores the last level the player played and the highest level they have gotten to. When the player gets to a new level the max level is increase which allows the player to load those levels from the load menu. If the player quits the application the progression game objects progression is serialised.